requirements.txt
Brotli==1.0.9
click==8.0.3
dash==2.0.0
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Compress==1.10.1
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
opencv-contrib-python==4.5.4.60
pandas==1.3.4
plotly==5.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.0.0
typing_extensions==4.0.0
Werkzeug==2.0.2
zipp==3.6.0
Pillow==8.4.0
